pyproject.toml
[tool.poetry]
name = "kcsec"
version = "0.1.0"
description = ""
authors = ["KevinCybura <KevinCybura@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django-extensions = "^3.0.5"
psycopg2-binary = "^2.8.5"
django-postgres-extra = "^1.22"
python-dotenv = "^0.14.0"
channels = "^2.4.0"
channels_redis = "^3.1.0"
django = "^3.1.1"
whitenoise = "^5.2.0"
django-storages = "^1.10"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-bootstrap4 = "^2.2.0"
websockets = "^8.1"
django-cors-headers = "^3.5.0"
django-redis = "^4.12.1"
simplejson = "^3.17.2"
Pillow = "^8.0.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
isort = "^5.5.1"
pytest = "^6.0.1"
black = "^20.8b1"
pytest-django = "^3.9.0"
coverage = "^5.2.1"
factory-boy = "^3.0.1"
selenium = "^3.141.0"
pylint = "^2.6.0"
django-debug-toolbar = "^3.2"

[tool.isort]
force_single_line = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.github
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | dist
    | staticfiles
  )/
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
